fase3/requirements.txt
pymongo[srv]
dnspython
python-dotenv
